+++ pkg/config/config_validator_helpers_test.go
@@ -1,0 +1,70 @@
+package config
+
+import (
+	"strings"
+	"testing"
+
+	"github.com/portpowered/infinite-you/pkg/interfaces"
+)
+
+type stubRequiredToolChecker map[string]RequiredToolCheckResult
+
+func (s stubRequiredToolChecker) Check(tool interfaces.RequiredToolConfig) RequiredToolCheckResult {
+	if result, ok := s[tool.Command]; ok {
+		return result
+	}
+	return RequiredToolCheckResult{}
+}
+
+func testBaseConfig() *interfaces.FactoryConfig {
+	return &interfaces.FactoryConfig{
+		WorkTypes: []interfaces.WorkTypeConfig{{
+			Name: "task",
+			States: []interfaces.StateConfig{
+				{Name: "init", Type: interfaces.StateTypeInitial},
+				{Name: "done", Type: interfaces.StateTypeTerminal},
+				{Name: "failed", Type: interfaces.StateTypeFailed},
+			},
+		}},
+		Workers: []interfaces.WorkerConfig{{Name: "w1"}},
+	}
+}
+
+func assertErrString(message string) error {
+	return &staticErr{message: message}
+}
+
+type staticErr struct {
+	message string
+}
+
+func (e *staticErr) Error() string {
+	return e.message
+}
+
+func assertFindingExists(t *testing.T, findings []Finding, rule string) {
+	t.Helper()
+	for _, f := range findings {
+		if f.Rule == rule && f.Severity == SeverityError {
+			return
+		}
+	}
+	t.Fatalf("expected error finding with rule %q, got %v", rule, findings)
+}
+
+func assertFindingMatch(t *testing.T, findings []Finding, rule string, pathSubstring string, messageSubstring string) {
+	t.Helper()
+	for _, f := range findings {
+		if f.Rule != rule || f.Severity != SeverityError {
+			continue
+		}
+		if !strings.Contains(f.Path, pathSubstring) {
+			t.Fatalf("finding path = %q, want substring %q", f.Path, pathSubstring)
+		}
+		if !strings.Contains(f.Message, messageSubstring) {
+			t.Fatalf("finding message = %q, want substring %q", f.Message, messageSubstring)
+		}
+		return
+	}
+	t.Fatalf("expected error finding with rule %q, got %v", rule, findings)
+}
